move tests to matching namespace
