+++ Raft.Test/Communication/CandidateNodeRunnerShould.cs
@@ -5,7 +5,7 @@ using Raft.Entities;
 using Raft.NodeStrategy;
 using Shouldly;
 
-namespace Raft.Test.Election
+namespace Raft.Test.Communication
 {
     [TestFixture]
     public class CandidateNodeRunnerShould
